Move failing action tests to post-commit command
The intent-to-add methods in the pre-commit beforeHook() and afterHook() caused coverage problems that I couldn't figure out. Moving these tests to use a different command (PostCommit instead of PreCommit) fixes the coverage issue because PostCommit doesn't have fancy features in its beforeHook() or afterHook() methods.
