+++ tests/CaptainHook/Console/Command/Hook/PostCommitTest.php
@@ -13,6 +13,7 @@ namespace CaptainHook\App\Console\Command\Hook;
 
 use CaptainHook\App\Console\Runtime\Resolver;
 use CaptainHook\App\Git\DummyRepo;
+use Exception;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\NullOutput;
 use PHPUnit\Framework\TestCase;
@@ -44,5 +45,48 @@ class PostCommitTest extends TestCase
         chdir($cwd);
 
         $this->assertTrue(true);
+    }
+
+    public function testExecuteFailingActionInDebugMode(): void
+    {
+        $this->expectException(Exception::class);
+
+        $output = $this->createMock(NullOutput::class);
+        $output->expects($this->once())->method('isDebug')->willReturn(true);
+
+        $resolver = new Resolver();
+        $repo     = new DummyRepo();
+        $input    = new ArrayInput(
+            [
+                '--configuration' => CH_PATH_FILES . '/config/valid-but-failing.json',
+                '--git-directory' => $repo->getGitDir(),
+                '--bootstrap'     => '../bootstrap/empty.php'
+            ]
+        );
+
+        $cmd = new PostCommit($resolver);
+        $cmd->run($input, $output);
+
+        $this->assertTrue(true);
+    }
+
+    public function testExecuteFailingActionInVerboseMode(): void
+    {
+        $output = $this->createMock(NullOutput::class);
+        $output->expects($this->once())->method('isDebug')->willReturn(false);
+        $output->expects($this->atLeast(1))->method('writeLn');
+
+        $resolver = new Resolver();
+        $repo     = new DummyRepo();
+        $input    = new ArrayInput(
+            [
+                '--configuration' => CH_PATH_FILES . '/config/valid-but-failing.json',
+                '--git-directory' => $repo->getGitDir(),
+                '--bootstrap'     => '../bootstrap/empty.php'
+            ]
+        );
+
+        $cmd = new PostCommit($resolver);
+        $this->assertEquals(1, $cmd->run($input, $output));
     }
 }
