+++ tests/CaptainHook/Console/Command/Hook/PreCommitTest.php
@@ -97,57 +97,4 @@ class PreCommitTest extends TestCase
 
         $this->assertTrue(true);
     }
-
-    /**
-     * Tests PreCommit::run
-     *
-     * @throws \Exception
-     */
-    public function testExecuteFailingActionInDebugMode(): void
-    {
-        $this->expectException(Exception::class);
-
-        $output = $this->createMock(NullOutput::class);
-        $output->expects($this->once())->method('isDebug')->willReturn(true);
-
-        $resolver = new Resolver();
-        $repo     = new DummyRepo();
-        $input    = new ArrayInput(
-            [
-                '--configuration' => CH_PATH_FILES . '/config/valid-but-failing.json',
-                '--git-directory' => $repo->getGitDir(),
-                '--bootstrap'     => '../bootstrap/empty.php'
-            ]
-        );
-
-        $cmd = new PreCommit($resolver);
-        $cmd->run($input, $output);
-
-        $this->assertTrue(true);
-    }
-
-    /**
-     * Tests PreCommit::run
-     *
-     * @throws \Exception
-     */
-    public function testExecuteFailingActionInVerboseMode(): void
-    {
-        $output = $this->createMock(NullOutput::class);
-        $output->expects($this->once())->method('isDebug')->willReturn(false);
-        $output->expects($this->atLeast(1))->method('writeLn');
-
-        $resolver = new Resolver();
-        $repo     = new DummyRepo();
-        $input    = new ArrayInput(
-            [
-                '--configuration' => CH_PATH_FILES . '/config/valid-but-failing.json',
-                '--git-directory' => $repo->getGitDir(),
-                '--bootstrap'     => '../bootstrap/empty.php'
-            ]
-        );
-
-        $cmd = new PreCommit($resolver);
-        $this->assertEquals(1, $cmd->run($input, $output));
-    }
 }
